requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# PDF parsing
pdfplumber==0.11.0

# Google Gemini + Embeddings
google-genai>=0.8.0

# FAISS semantic search
faiss-cpu==1.7.4
numpy>=1.26.0